Extended payload length bug. #229

+++ source/NetCoreServer/UdpClient.cs
@@ -457,7 +457,7 @@ namespace NetCoreServer
             try
             {
                 // Sent datagram to the server
-                int sent = Socket.SendTo(buffer, SocketFlags.None, endpoint);
+                long sent = Socket.SendTo(buffer, SocketFlags.None, endpoint);
                 if (sent > 0)
                 {
                     // Update statistic
@@ -632,7 +632,7 @@ namespace NetCoreServer
             try
             {
                 // Receive datagram from the server
-                int received = Socket.ReceiveFrom(buffer, (int)offset, (int)size, SocketFlags.None, ref endpoint);
+                long received = Socket.ReceiveFrom(buffer, (int)offset, (int)size, SocketFlags.None, ref endpoint);
 
                 // Update statistic
                 DatagramsReceived++;
